requirements.txt
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyMuPDF
python-docx
pytesseract
Pillow
spacy
Flask
requests
openai
mysql-connector-python
bcrypt
pyjwt
gunicorn
pymysql